Add some stuff

+++ DataStructures/IHeapItem.cs
@@ -1,6 +1,6 @@
 namespace IDontWantToDoThisTwice.DataStructures {
 
 public interface IHeapItem {
-    float Priority { get; }
+    double Priority { get; }
 }
 }
